+++ library/src/main/jni/com_twilio_video_LocalAudioTrack.cpp
@@ -36,6 +36,13 @@ std::shared_ptr<twilio::media::LocalAudioTrack> getLocalAudioTrack(jlong local_a
     return local_audio_track_context->getLocalAudioTrack();
 }
 
+std::string getLocalAudioTrackHash(std::shared_ptr<twilio::media::LocalAudioTrack> local_audio_track) {
+    std::size_t local_audio_track_hash =
+            std::hash<twilio::media::LocalAudioTrack *>{}(local_audio_track.get());
+
+    return std::to_string(local_audio_track_hash);
+}
+
 jobject createJavaLocalAudioTrack(jobject j_context,
                                   std::shared_ptr<twilio::media::LocalAudioTrack> local_audio_track) {
     JNIEnv *jni = webrtc_jni::GetEnv();
@@ -48,10 +55,12 @@ jobject createJavaLocalAudioTrack(jobject j_context,
     LocalAudioTrackContext* local_audio_track_context =
             new LocalAudioTrackContext(local_audio_track);
     jstring j_name = JavaUTF16StringFromStdString(jni, local_audio_track->getName());
+    jstring j_track_hash = JavaUTF16StringFromStdString(jni,
+                                                        getLocalAudioTrackHash(local_audio_track));
     jobject j_local_audio_track = jni->NewObject(j_local_audio_track_class,
                                                  j_local_audio_track_ctor_id,
                                                  webrtc_jni::jlongFromPointer(local_audio_track_context),
-                                                 JavaUTF16StringFromStdString(jni, local_audio_track->getTrackId()),
+                                                 j_track_hash,
                                                  j_name,
                                                  local_audio_track->isEnabled(),
                                                  j_context);